+++ spacy_conll/__main__.py
@@ -10,6 +10,8 @@ def main():
     parser.add_argument('-a', '--input_encoding', default=getpreferredencoding(),
                         help="Encoding of the input file. Default value is system default.")
     parser.add_argument('-b', '--input_str', default=None, help="Input string to parse.")
+    parser.add_argument('-t', '--is_tokenized', default=False, action="store_true",
+                        help="Enable this option when your text has already been tokenized (space-seperated).")
     # Output arguments
     parser.add_argument('-o', '--output_file', default=None,
                         help="Path to output file. If not specified, the output will be printed on standard output.")
@@ -20,11 +22,11 @@ def main():
     parser.add_argument('-n', '--nlp', default=None,
                         help="Optional already initialised spaCy NLP model. Has precedence over 'model'.")
     # Additional arguments
-    parser.add_argument('-v', '--verbose', default=False, action="store_true",
-                        help="To print the output to stdout, regardless of 'output_file'.")
     parser.add_argument('-d', '--include_headers', default=False, action="store_true",
                         help="To include headers before every sentence's output. These headers include the sentence"
                              " text and the sentence ID.")
+    parser.add_argument('-v', '--verbose', default=False, action="store_true",
+                        help="To print the output to stdout, regardless of 'output_file'.")
 
     args = parser.parse_args()