+++ HISTORY.rst
@@ -3,6 +3,12 @@ History
 #######
 
 **************************
+0.1.0 (January 16th, 2019)
+**************************
+Added the :code:`-t|--is_tokenized` flag. It indicates that your text has already been tokenized and that it should not
+be tokenized again. The tokens should be space-separated, e.g. :code:`I like grandma 's cookies !`.
+
+**************************
 0.0.3 (January 14th, 2019)
 **************************
 Initial commit.